in OvOp: used TryParse in Input, replaced Coordinates3D to a new file, deleted Print, overloaded ToString for class Coordinates3D
